00-original/sources/py/srcV2/requirements.txt
lxml==4.5.1
MarkupSafe==1.1.1
mysql-connector-python==8.0.20
numpy==1.18.4
pandas==1.0.4
protobuf==3.12.2
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
six==1.15.0
